ai_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1

# Embeddings & Vector Search - Compatible versions
sentence-transformers>=2.2.0
faiss-cpu>=1.9.0
numpy>=1.26.0

# Optional: Gemini Integration
google-generativeai>=0.3.0

# Utilities
python-multipart==0.0.6
